+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - New #383: Add `caseSensitive` option to like condition (@vjik)
 - Enh #386: Remove `getCacheKey()` and `getCacheTag()` methods from `Schema` class (@Tigrov)
 - Bug #388: Set empty `comment` and `extra` properties to `null` when loading table columns (@Tigrov)
+- Enh #389, #390: Use `DbArrayHelper::arrange()` instead of `DbArrayHelper::index()` method (@Tigrov)
 
 ## 1.2.0 March 21, 2024